Refactor: Split monolithic init.lua into focused modules
Split 272-line init.lua into 4 modules for better maintainability:

- utils.lua (50 lines): Utility functions
  - debug_log, safe_call, parse_word, determine_henkan_type

- skkeleton.lua (49 lines): Skkeleton/denops integration
  - is_enabled, get_completion_data, register_completion

- completion.lua (133 lines): Completion item builder
  - convert_ranks_to_map, build_text_edit_range
  - build_completion_item, build_completion_items

- init.lua (107 lines): Main source API
  - Implements blink.cmp source interface
  - Orchestrates other modules

Benefits:
- Clear separation of concerns
- Easier to understand and maintain
- Better testability (all 11 tests pass)
- Reduced cognitive load per file
